MetricsGenerator/requirements.txt
kafka-python
requests
minio
python-dotenv
pandas